Build fat, shaded jar for custom external dependencies
During test it became clear that several dependencies could not simply
be removed, but actually needed to be shaded to deal with
incompabilities with spark.

This PR exports the custom dependencies as a single, fat jar, and adds
shading for `com.jackson` databinding (which has incompatible versions
between the AWS SDK and Spark).